% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_model_list}
\alias{apply_model_list}
\title{Compute likelihood of many models on the same data}
\usage{
apply_model_list(data, model_list, lhood_fun)
}
\arguments{
\item{data}{data.frame with columns F1 and F2 (passed to marginal_model_lhood)}

\item{model_list}{list of models to calculate likelihood}

\item{lhood_fun}{likelihood function}
}
\value{
data.frame of likelihoods, with one column per model, one row per row
in data
}
\description{
Compute likelihood of many models on the same data
}
