% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\alias{classify_vowels}
\title{Use trained models to classify observed formant values}
\usage{
classify(data, models, category)

classify_vowels(data, models)
}
\arguments{
\item{data}{Data frame with grouping columns used to generate models
and F1,F2}

\item{models}{Data frame with column \code{category} and list column `model`
(as generated by `train_models`).}

\item{category}{Name of the column in \code{models} that has category labels}
}
\value{
Data frame with one row per data x model combination, with likelihood
  in lhood, posterior probability in posterior. The model with the highest
  posterior for each token has TRUE in posterior_choice.
}
\description{
Names of columns are taken from the names of the first \code{mu} in models.
}
\section{Functions}{
\itemize{
\item \code{classify_vowels}: Special case where \code{category = "Vowel"}
}}

