% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_indexical_with_holdout}
\alias{classify_indexical_with_holdout}
\title{Classify test data with cross-validated models}
\usage{
classify_indexical_with_holdout(data_and_models)
}
\arguments{
\item{data_and_models}{output of train_models_indexical_with_holdout, with
one row per talker (or holdout grouping level), and list columns for
\code{data_test} and \code{models}.}
}
\value{
Data frame with one row per combination of data row and model, with
  columns corresponding to the held-out and grouping variables, plus model,
  lhood (total log-likelihood of data under model), log_posterior, posterior,
  and posterior_choice (1 for the MAP model and 0 otherwise)
}
\description{
See \code{\link{train_models_indexical_with_holdout}}.
}
