% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{covar2D}
\alias{covar2D}
\title{make a 2D variance-covariance matrix from a vector of vars and covar}
\usage{
covar2D(dim1, dim2, cov12)
}
\arguments{
\item{dim1}{variance of dimension 1}

\item{dim2}{variance of dimension 2}

\item{cov12}{covariance}
}
\description{
make a 2D variance-covariance matrix from a vector of vars and covar
}
