% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cue_matrix}
\alias{cue_matrix}
\title{Extract matrix of cues from data frame columns}
\usage{
cue_matrix(d, cues)
}
\arguments{
\item{d}{data frame}

\item{cues}{quoted names of columns to pull out}
}
\value{
matrix with values from named columns
}
\description{
Extract matrix of cues from data frame columns
}
