% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{formants_matrix}
\alias{formants_matrix}
\title{Extract matrix of formants from data frame columns}
\usage{
formants_matrix(d, formants = c("F1", "F2"))
}
\arguments{
\item{d}{data frame}

\item{formants}{=c('F1', 'F2') quoted names of columns to pull out}
}
\value{
matrix with values from named columns
}
\description{
Extract matrix of formants from data frame columns
}
