% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Two methods to compute accuracy from classifier}
\usage{
get_accuracy(tbl, category_col, method = "choice")
}
\arguments{
\item{tbl}{classifier output}

\item{category_col}{name of column specifying true category}

\item{method}{if 'choice' (default), accuracy is 1 if choice is correct,
0 if not. if 'posterior', accuracy is posterior probability of true
category}
}
\value{
tbl with accuracy in column \code{accuracy}
}
\description{
Two methods to compute accuracy from classifier
}
