% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{KL.bvnorm}
\alias{KL.bvnorm}
\title{Calculate the KL divergence between two bivariate Gaussians}
\usage{
KL.bvnorm(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1, mu2}{mean vector of two distributions}

\item{sigma1, sigma2}{covariance matrix in lower triangular vector (see
\code{\link{covar2D}}) or matrix form}
}
\description{
Calculate the KL divergence between two bivariate Gaussians
}
