% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{list_models}
\alias{list_models}
\title{Convert data frame of models to named list}
\usage{
list_models(d, names_col = NULL, model_field = "model")
}
\arguments{
\item{d}{data.frame of models}

\item{names_col}{(optional) name of column to be used for names of models}

\item{model_field}{='model' name of column with models}
}
\value{
a named list of models
}
\description{
Convert data frame of models to named list
}
