% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{lower_tri_to_full}
\alias{lower_tri_to_full}
\title{Expand lower-triangular vector to full matrix}
\usage{
lower_tri_to_full(lower_tri_vec)
}
\arguments{
\item{lower_tri_vec}{Lower triangular vector of length p*(p+1)/2}

\item{return}{Symmetric, square p×p matrix}
}
\description{
Expand lower-triangular vector to full matrix
}
