% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{marginal_model_lhood}
\alias{marginal_model_lhood}
\title{Marginal likelihood of data under mixture model}
\usage{
marginal_model_lhood(mods, data, log = TRUE, ...)
}
\arguments{
\item{mods}{list of models in mixture.}

\item{data}{matrix with observations in rows and dimensions in columns}

\item{log}{=TRUE returns log likelihood (default)}

\item{...}{additional arguments passed nspvowels::model_lhood (mvtnorm::dmvnorm)}
}
\value{
vector with marginal likelihood for each row in data.
}
\description{
Calls \code{\link{model_lhood}} on each model and takes the average (assumes
equal prior/mixing weights)
}
