% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{model_lhood}
\alias{model_lhood}
\title{Likelihood of data under one vowel's model}
\usage{
model_lhood(mod, dat, ...)
}
\arguments{
\item{mod}{multivariate normal model (list with mean vector mu and covariance
matrix Sigma)}

\item{dat}{matrix with observations in rows and dimensions in columns, passed to
mvtnorm::mvnorm.}

\item{...}{additional arguments are passed to dmvnorm}
}
\description{
Likelihood of data under one vowel's model
}
