% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{model_matrix}
\alias{model_matrix}
\title{Create data matrix for a model}
\usage{
model_matrix(model, data)
}
\arguments{
\item{model}{a model (with field \code{mu}, a named vector) or a list of
models that will be flattened until there's a field named \code{mu}.}

\item{data}{a data frame with columns of the same names as the entries in
model's \code{mu}}
}
\value{
a matrix with observations in rows and dimensions in columns.
}
\description{
Extracts columns from data frame with names from model's \code{mu}
field. Operates recursively, so it will work on a list of models
}
