% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sample_n_groups}
\alias{sample_n_groups}
\title{Randomly subsample data by group}
\usage{
sample_n_groups(tbl, group, n, ...)
}
\arguments{
\item{tbl}{tbl of data}

\item{group}{(quoted) grouping variable}

\item{n}{number of samples}

\item{...}{additional arguments are passed to dplyr::sample_n.}
}
\value{
A tbl with \code{n} levels of \code{group} randomly sampled within
  each existing grouping level of \code{tbl}. Any grouping of \code{tbl} is
  preserved in the return value.
}
\description{
Wraps dplyr::sample_n.  Subsampling will be done at the level of the
specified \code{group} column, \emph{within} any grouping that already exists
in the input tbl.
}
