% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_models_indexical_with_holdout}
\alias{train_models_indexical_with_holdout}
\title{Train indexical models with cross-validation}
\usage{
train_models_indexical_with_holdout(d, groups, category = "Vowel",
  holdout = "Talker", ...)
}
\arguments{
\item{d}{data frame (a la `nsp_vows`)}

\item{groups}{quoted name of indexical grouping variable column (e.g.,
'Dialect'). One model will be created for each level of this variable.}

\item{category}{='Vowel' quoted name of column for linguistic category. Each
indexical group's model is a list of individual Vowel models}

\item{holdout}{='Talker' unit to perform cross-validation on. one row per
level of this variable is created with models trained after removing the
corresponding level.}

\item{...}{additional arguments are passed to \code{\link{train_models}}}
}
\value{
the dataframe returned by \code{\link{train_test_split}}, plus a
  models list column, each entry of which is a model for one level of
  \code{groups} after holding out that row's Talker (or level of holdout).
}
\description{
Combines \code{\link{train_test_split}}, \code{\link{train_models}}, and
\code{\link{list_models}} to produce a list of indexical groups' models.
}
\details{
Each group's indexical model is a mixture of models at the linguistic level
(defaults to Vowel)
}
