% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_test_split}
\alias{train_test_split}
\title{Split data into training and test sets}
\usage{
train_test_split(d, holdout, groups = NULL)
}
\arguments{
\item{d}{data frame}

\item{holdout}{quoted name of column that defines train/test splits.}

\item{groups}{quoted name of column(s) to group training data by. optional,
will be preserved in output (and added as grouping to training data).}
}
\value{
data frame with columns `data_train`, `data_test`, one for each level
  of holdout and group (if specified), which are also included. `data_test`
  has the corresponding subset of the input data, and `data_train` has the
  rest.
}
\description{
Split data into training and test sets
}
