% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{KL_mvnorm}
\alias{KL_mvnorm}
\title{Calculate the KL divergence between two multivariate Gaussians}
\usage{
KL_mvnorm(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1, mu2}{mean vector of two distributions}

\item{sigma1, sigma2}{covariance matrix in lower triangular vector (see
\code{\link{lower_tri_to_full}}) or matrix form}
}
\description{
Calculate the KL divergence between two multivariate Gaussians
}
