% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_models}
\alias{train_models}
\title{Train models on specified grouping variable.}
\usage{
train_models(data, grouping = "Vowel", formants = c("F1", "F2"),
  add_groups = TRUE)
}
\arguments{
\item{data}{Data source with columns for grouping and "formants"}

\item{grouping}{quoted column names to group by (strings, `quote`, or `~`)
(default "Vowels")}

\item{formants}{quoted column names containing formants (or other measures).
(default: c("F1", "F2"))}

\item{add_groups}{add `grouping` to existing groups, or overwrite (default TRUE)}
}
\value{
A data frame like for summarise, with columns for grouping factor, plus
  list columns `data` (the matrix used to train the model) and `model`. `model$mu`
  is the mean, and `model$Sigma` is the covariance matrix.  If `add_groups==TRUE`,
  original grouping will be maintained.
}
\description{
Grouping defaults to Vowels, and will be added to any grouping already
present by default.
}
