% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_models_indexical_subsample_holdout}
\alias{train_models_indexical_subsample_holdout}
\title{Train indexical group models with holdout and subsampling talkers}
\usage{
train_models_indexical_subsample_holdout(d, groups, n_subsample,
  category = "Vowel", holdout = "Talker")
}
\arguments{
\item{d}{data frame (a la `nsp_vows`)}

\item{groups}{quoted name of indexical grouping variable column (e.g.,
'Dialect'). One model will be created for each level of this variable.}

\item{n_subsample}{Number of holdout levels to subsample for training data.}

\item{category}{='Vowel' quoted name of column for linguistic category. Each
indexical group's model is a list of individual Vowel models}

\item{holdout}{='Talker' unit to perform cross-validation on. one row per
level of this variable is created with models trained after removing the
corresponding level.}
}
\value{
the dataframe returned by \code{\link{train_test_split}}, with
  data_train replaced by the subsampled version for each talker, plus a
  models list column, each entry of which is a model for one level of
  \code{groups} after holding out that row's Talker (or level of holdout).
}
\description{
Like \code{\link{train_models_indexical_with_holdout}}, but randomly
sub-samples training data to have specified number of talkers.
}
